# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AxisFitOut": {
        "description": "Why one axis scored what it did \u2014 Wave 4 renders the band, not the number.",
        "properties": {
          "band": {
            "title": "Band",
            "type": "string"
          },
          "half_widths_away": {
            "title": "Half Widths Away",
            "type": "number"
          },
          "niche_width": {
            "title": "Niche Width",
            "type": "number"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "species_value": {
            "title": "Species Value",
            "type": "number"
          },
          "width_estimated": {
            "title": "Width Estimated",
            "type": "boolean"
          }
        },
        "required": [
          "band",
          "score",
          "half_widths_away",
          "species_value",
          "niche_width",
          "width_estimated"
        ],
        "title": "AxisFitOut",
        "type": "object"
      },
...
      "BedSuggestions": {
        "description": "Suggestions for one bed, with the site vector they were ranked against \u2014\nso the UI can say what it matched on rather than showing a bare list.",
        "properties": {
          "bed_id": {
            "title": "Bed Id",
            "type": "integer"
          },
          "bed_name": {
            "title": "Bed Name",
            "type": "string"
          },
          "filters": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FilterCountsOut"
            },
            "default": {},
            "title": "Filters",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlantSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "light_state": {
            "enum": [
              "missing",
              "stale",
              "current"
            ],
            "title": "Light State",
            "type": "string"
          },
          "site_axes": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Site Axes",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "woody": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlantSummary"
            },
            "title": "Woody",
            "type": "array"
          },
          "woody_total": {
            "default": 0,
            "title": "Woody Total",
            "type": "integer"
          }
        },
        "required": [
          "bed_id",
          "bed_name",
          "site_axes",
          "total",
          "items",
          "light_state"
        ],
        "title": "BedSuggestions",
        "type": "object"
      },
...
      "FilterCountsOut": {
        "description": "How one active filter divided the candidate set.\n\nReported so the UI can say what was left out. `unknown` is not a rounding\nerror: height is recorded for 44% of German species and colour for 6.6%, and\na filter that hides that is indistinguishable from one that is broken.",
        "properties": {
          "excluded": {
            "title": "Excluded",
            "type": "integer"
          },
          "matched": {
            "title": "Matched",
            "type": "integer"
          },
          "unknown": {
            "title": "Unknown",
            "type": "integer"
          }
        },
        "required": [
          "matched",
          "unknown",
          "excluded"
        ],
        "title": "FilterCountsOut",
        "type": "object"
      },
      "FitOut": {
        "properties": {
          "axes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AxisFitOut"
            },
            "title": "Axes",
            "type": "object"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "score",
          "axes"
        ],
        "title": "FitOut",
        "type": "object"
      },
...
      "GardenCreate": {
        "description": "Creating a garden. Latitude and longitude are range-checked here, before\nthey reach solar code that would happily compute a sun path for latitude 500.",
        "properties": {
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "latitude",
          "longitude"
        ],
        "title": "GardenCreate",
        "type": "object"
      },
      "GardenCreated": {
        "description": "The token is returned; the numeric id deliberately is not.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "share_token": {
            "title": "Share Token",
            "type": "string"
          }
        },
        "required": [
          "share_token",
          "name"
        ],
        "title": "GardenCreated",
        "type": "object"
      },
...
      "GrowthForm": {
        "description": "The growth forms the catalogue actually records.\n\nA closed set rather than a free string. The value never reaches SQL, but an\nunbounded parameter that silently matches nothing is its own kind of lie \u2014\nthe user cannot tell a typo from an empty catalogue.",
        "enum": [
          "forb",
          "herb",
          "graminoid",
          "shrub",
          "subshrub",
          "tree"
        ],
        "title": "GrowthForm",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlantSummary": {
        "properties": {
          "bird_partners": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bird Partners"
          },
          "canonical_name": {
            "title": "Canonical Name",
            "type": "string"
          },
          "colour_known": {
            "title": "Colour Known",
            "type": "boolean"
          },
          "family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family"
          },
          "fit": {
            "$ref": "#/components/schemas/FitOut"
          },
          "fits_bed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fits Bed"
          },
          "flower_colour": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flower Colour"
          },
          "flowering_end_month": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flowering End Month"
          },
          "flowering_start_month": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flowering Start Month"
          },
          "height_max_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Max M"
          },
          "observed_colour": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Colour"
          },
          "space_m2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space M2"
          },
          "taxon_id": {
            "title": "Taxon Id",
            "type": "integer"
          }
        },
        "required": [
          "taxon_id",
          "canonical_name",
          "family",
          "height_max_m",
          "flowering_start_month",
          "flowering_end_month",
          "flower_colour",
          "colour_known",
          "bird_partners",
          "space_m2",
          "fits_bed",
          "fit"
        ],
        "title": "PlantSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NinaNatur",
    "version": "contract"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/gardens": {
      "post": {
        "description": "Create a garden. The token comes back here and is the only way back in.\n\nA signed-in visitor's garden is theirs from the moment it exists. It used to\nbe nobody's unless something called `/claim` afterwards, and nothing did \u2014\nso an account kept the links and then listed none of them.\n\n`current_account` rather than `require_account`: making a garden without an\naccount is the ordinary case, and the token is still the whole of its\naccess control either way.",
        "operationId": "create_api_v1_gardens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GardenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GardenCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "gardens"
        ]
      }
    },
...
    "/api/v1/gardens/{token}/beds/{bed_id}/suggestions": {
      "get": {
        "description": "Species that suit this bed, ranked by fit against its own site vector.\n\nWoody plants get a shortlist of their own. Introduced species are left out\n(the product promises native plants), and so, unless asked for, are species\nthe bed is far too bright for (`filters.light_verdict`). `light_state` says\nwhether there was a light value to judge by.",
        "operationId": "bed_suggestions_api_v1_gardens__token__beds__bed_id__suggestions_get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bed_id",
            "required": true,
            "schema": {
              "title": "Bed Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "colour",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Colour"
            }
          },
          {
            "in": "query",
            "name": "height_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Height Min"
            }
          },
          {
            "in": "query",
            "name": "height_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Height Max"
            }
          },
          {
            "in": "query",
            "name": "flowering_month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 12,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flowering Month"
            }
          },
          {
            "in": "query",
            "name": "growth_form",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/GrowthForm"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Growth Form"
            }
          },
          {
            "in": "query",
            "name": "include_unknown",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unknown",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_trees",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Trees",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_introduced",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Introduced",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "exclude_planted",
            "required": false,
            "schema": {
              "default": true,
              "title": "Exclude Planted",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_light_unsuitable",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Light Unsuitable",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BedSuggestions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bed Suggestions",
        "tags": [
          "planning"
        ]
      }
    },
...
  }
}